test: move tests to test folder
