+++ test/greet.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-option go_package = "github.com/alexfalkowski/go-service/pkg/grpc/internal";
+option go_package = "github.com/alexfalkowski/go-service/test";
 
 import "google/api/annotations.proto";
