+++ tools/protoc
@@ -3,7 +3,7 @@
 set -e
 
 # Generate for testing.
-protoc -I pkg/grpc/internal --go_out=pkg/grpc/internal --go_opt=paths=source_relative \
-    --go-grpc_out=pkg/grpc/internal --go-grpc_opt=paths=source_relative \
-    --grpc-gateway_out=pkg/grpc/internal --grpc-gateway_opt=paths=source_relative \
-    pkg/grpc/internal/greet.proto
+protoc -I test --go_out=test --go_opt=paths=source_relative \
+    --go-grpc_out=test --go-grpc_opt=paths=source_relative \
+    --grpc-gateway_out=test --grpc-gateway_opt=paths=source_relative \
+    test/greet.proto
